refactor: 공통 스타일 컴포넌트 분리
