+++ src/pages/member/[detail].tsx
@@ -12,6 +12,7 @@ import { memberApi } from "@/api/memberApi";
 import Loading from "@/components/common/loading/Loading";
 import { GetServerSideProps } from "next";
 import { parseCookies } from "@/util/parseCookie";
+import { PageContainer } from "@/styles/common.style";
 
 const flag = 0;
 
@@ -95,13 +96,8 @@ const Member = () => {
 
 export default Member;
 
-const Container = styled.div`
-  width: 100vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  background-color: ${(props) => props.theme.backgronudColors.white};
+const Container = styled(PageContainer)`
+
 `;
 
 const InfoContainer = styled.div`
