+++ src/pages/study/[studyId]/index.tsx
@@ -12,6 +12,7 @@ import Loading from "@/components/common/loading/Loading";
 import { parseCookies } from "@/util/parseCookie";
 import { GetServerSideProps } from "next";
 import { useState, useEffect } from "react";
+import { PageContainer } from "@/styles/common.style";
 
 interface IServerSideProp {
   refreshToken: string;
@@ -111,14 +112,7 @@ const StudyDetail = ({ refreshToken, memberId }: IServerSideProp) => {
 
 export default StudyDetail;
 
-const Container = styled.div`
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  background-color: ${(props) => props.theme.backgronudColors.white};
+const Container = styled(PageContainer)`
 `;
 
 const StudyContainer = styled.div`
