+++ src/styles/common.style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const PageContainer = styled.div`
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  background-color: ${(props) => props.theme.backgronudColors.white};
+`;
